Arrange classes in separate files.
