+++ src/base/colors/color-calculator.ts
@@ -1,0 +1,11 @@
+import { IColorHelper } from './i-color-helper';
+import { BaseAccessor } from '../../core/types';
+
+export class ColorCalculator implements IColorHelper {
+    public colorAccessor: BaseAccessor<string>;
+    public getColor: BaseAccessor<string>;
+
+    constructor(colorCalculator: BaseAccessor<string>) {
+        this.getColor = colorCalculator;
+    }
+}
