+++ src/base/colors/i-color-helper.ts
@@ -1,0 +1,6 @@
+import { BaseAccessor } from '../../core/types';
+
+export interface IColorHelper {
+    getColor(d, i?: number): string;
+    colorAccessor: BaseAccessor<string>;
+}
